Added debug logging (for myself). Updated the raw byte values. Increased polling rate to 2ms
The multi-controller support seems to work based on Linux logs, but it currently crashes Steam. There might be an ID conflict with the virtual controllers.

The log statements shouldn't affect the performance of this driver when the log level isn't set to DEBUG. JIT should be able to skip past all the logging framework's if-checks